move "new field/generate UUID" tests to discrete module
